Refactor: Reorganize codebase into modular file structure
Major restructuring to improve maintainability and code organization:

## File Structure Changes
- Split monolithic 652-line ExerciseViews.swift into focused modules
- Organized code into Models/, Views/, Services/ directories
- Created reusable UI components in Views/Components/

## New Files
- Models/ExerciseModels.swift - All data models and types
- Services/MusicTheoryData.swift - Centralized music theory data & config
- Views/Components/SharedComponents.swift - Reusable UI components
- Views/ExerciseViews/ - Individual exercise view files (7 files)

## Improvements
- Eliminated code duplication with shared components
- Centralized configuration and data constants
- Improved separation of concerns
- Reduced individual file complexity
- Enhanced maintainability and testability
