+++ Views/ExerciseViews/PlaceholderViews.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+
+// MARK: - Placeholder Views for Future Implementation
+
+struct GoodrickTwoStringVampView: View {
+    var body: some View {
+        ComingSoonView(title: "Goodrick Two String Vamp", icon: "guitars.fill")
+    }
+}
+
+struct RandomExerciseView: View {
+    var body: some View {
+        ComingSoonView(title: "Random Exercise", icon: "shuffle")
+    }
+}
